Add fixes for getting setup to upgrade script

+++ upgrade.sh
@@ -91,10 +91,15 @@ fi
 
 info " ------ Starting FirstRun Upgrade ------"
 info "Getting latest for 'setupopenflixr'"
-if [[ -d /opt/OpenFLIXR2.SetupScript/ ]]; then
-    rm -r /opt/OpenFLIXR2.SetupScript/
+if [[ ! -d /opt/OpenFLIXR2.SetupScript/.git ]] && [[ ! -d /opt/OpenFLIXR2.SetupScript/.scripts ]]; then
+    if [[ -d /opt/OpenFLIXR2.SetupScript ]]; then
+        rm -r /opt/OpenFLIXR2.SetupScript/
+    fi
+    git clone https://github.com/openflixr/OpenFLIXR2.SetupScript /opt/OpenFLIXR2.SetupScript
+    chown -R openflixr:openflixr /opt/OpenFLIXR2.SetupScript || fatal "Failed to set permissions on '/opt/OpenFLIXR2.SetupScript'"
+else
+    chown -R openflixr:openflixr /opt/OpenFLIXR2.SetupScript || fatal "Failed to set permissions on '/opt/OpenFLIXR2.SetupScript'"
 fi
-git clone https://github.com/openflixr/OpenFLIXR2.SetupScript /opt/OpenFLIXR2.SetupScript
 
 if [[ -d /opt/OpenFLIXR2.SetupScript/.git ]] && [[ -d /opt/OpenFLIXR2.SetupScript/.scripts ]]; then
     cd "/opt/OpenFLIXR2.SetupScript/" || fatal "Failed to change to '/opt/OpenFLIXR2.SetupScript/' directory."
